fixes, not finished..

+++ VideoReading/getVideoTimeline.m
@@ -1,7 +1,7 @@
 function Timeline=getVideoTimeline(firstFrame, lastFrame, zeroFrame, triggerDelay, framesPerMillisecond)
 
-timeOffset=(zeroFrame-firstFrame)/framesPerMillisecond-triggerDelay;
-timeSize = lastFrame-firstFrame;
+timeOffset=double(zeroFrame-firstFrame)/framesPerMillisecond-triggerDelay;
+timeSize = double(lastFrame-firstFrame);
 Timeline= linspace(0,timeSize/framesPerMillisecond,timeSize)+timeOffset;
 
 end
